add tests and edit docstring

+++ src/image_recommender/metrics/embedding_distance.py
@@ -5,7 +5,7 @@ def cosine_distance(a: np.ndarray, b: np.ndarray) -> float:
     """
     Computes cosine distance between two 1D vectors.
     Input: a, b - 1D numpy arrays of the same shape.
-    Output: scalar float in [0, 2], where 0 means identical and 2 means opposite.
+    Output: scalar np.float64 in [0, 2], where 0 means identical and 2 means opposite.
     Possible error: ValueError if shapes differ or if either vector is zero.
     """
 
@@ -36,7 +36,7 @@ def cosine_distance_to_many(q: np.ndarray, X: np.ndarray) -> np.ndarray:
     Computes cosine distance between a single query vector and
     multiple vectors in a matrix.
     Input: q (1D numpy array of shape (D,)), X (2D numpy array of shape (N, D)).
-    Output: 1D numpy array of shape (N,) with cosine distances.
+    Output: 1D numpy array of shape (N,) as np.float64 with cosine distances.
     Possible error: ValueError if shapes differ or if either vector is zero.
     """
